fixing controller after merge

+++ src/quadruped_utils/include/quadruped_utils/ZeroJointsController.hpp
@@ -45,8 +45,8 @@ protected:
     std::vector<float> measured_state_values_;
 
     double zero_effort_lim;
-    float kp;
-    float kd;
+    double kp;
+    double kd;
 
     // Command interface vectors
     std::vector<std::reference_wrapper<hardware_interface::LoanedCommandInterface>>
